+++ CHANGELOG.md
@@ -8,6 +8,8 @@ NO RELEASE DATE SET YET
 
 #### Features
 * `EMP-10419` *Exposure* specific `OfflineFairplayRequester`
+* `EMP-9697` Carousel integration `FetchCarouselList`
+* Added `CustomerConfig` model
 
 #### Changes
 * Moved `FairplayError` from `Player` to `Exposure`.
@@ -16,6 +18,7 @@ NO RELEASE DATE SET YET
 * `EMP-10270` `Airing` is now distinct from fetching `Program`
 * `EMP-10322` `Image.Orientation` made `Equatable`.
 * Integrated `Download` entitlement requests and validation.
+* Renamed `Exposure` protocol to `ExposureType`
 
 #### Bug fixes
 * `EMP-10269` `FetchEPG` now respects the *onlyPublished* parameter. Fixed url encoding.